+++ libs/utils/source/ftrack_utils/actions/__init__.py
@@ -1,0 +1,7 @@
+# :coding: utf-8
+# :copyright: Copyright (c) 2024 ftrack
+
+from ftrack_utils.actions.remote_actions import (
+    default_discover_action_callback,
+    register_remote_action,
+)